+++ src/screens/surrealist/pages/Provision/style.module.scss
@@ -11,3 +11,18 @@
 		padding-block: 4px !important;
 	}
 }
+
+.provision-loader::after {
+	border: solid 3px transparent;
+	background-image: linear-gradient(var(--mantine-color-body), var(--mantine-color-body)), var(--surrealist-gradient);
+	background-origin: border-box;
+	background-clip: content-box, border-box;
+	animation-duration: 0.75s;
+}
+
+.provision-icon {
+	z-index: 0;
+	width: 68px;
+	height: 68px;
+	transform: translateY(1px);
+}
